content-formatting module
