+++ .storybook/components/TextField/TextField.stories.tsx
@@ -6,7 +6,7 @@ import {
 } from "@components/TextFields"
 import { ComponentMeta, ComponentStory } from "@storybook/react-native"
 import { Button, KeyboardAvoidingView, View } from "react-native"
-import { ContentText } from "@components/ContentText"
+import { ContentText } from "@content-formatting"
 import { Ionicon } from "@components/common/Icons"
 
 const TextFieldMeta: ComponentMeta<typeof TextField> = {
@@ -78,7 +78,13 @@ const ErrorView = () => {
       placeholder="Enter Text"
       onChangeText={setText}
       style={{ padding: 16 }}
-      error={<ContentText text="@die" onUserHandleTapped={() => {}} />}
+      error={
+        <ContentText
+          text="@die"
+          onUserHandleTapped={() => {}}
+          onEventHandleTapped={() => {}}
+        />
+      }
     />
   )
 }
@@ -89,7 +95,6 @@ const Password = () => {
   return (
     <View style={{ width: "100%" }}>
       <PasswordTextField
-        isFocused={isFocused}
         value={text}
         placeholder="Enter Text"
         onChangeText={setText}
