+++ auth/sign-up/ChangeUserHandle.test.tsx
@@ -1,11 +1,11 @@
 import { renderHook, waitFor } from "@testing-library/react-native"
 import { useSignUpChangeUserHandleForm } from "./ChangeUserHandle"
-import { UserHandle } from "@lib/users"
 import { act } from "react-test-renderer"
 import { fakeTimers } from "../../tests/helpers/Timers"
 import { TestQueryClientProvider } from "../../tests/helpers/ReactQuery"
 import { neverPromise } from "../../tests/helpers/Promise"
 import { captureAlerts } from "../../tests/helpers/Alerts"
+import { UserHandle } from "@content-formatting"
 
 describe("SignUpChangeUserHandle tests", () => {
   beforeEach(() => jest.resetAllMocks())
