+++ auth/sign-up/ChangeUserHandle.tsx
@@ -1,10 +1,10 @@
 import { AuthFormView } from "@auth/AuthLayout"
 import { AuthShadedTextField } from "@auth/AuthTextFields"
 import { Ionicon } from "@components/common/Icons"
+import { UserHandle, UserHandleError } from "@content-formatting"
 import { useFormSubmission } from "@hooks/FormHooks"
 import { sleep } from "@lib/DelayData"
 import { QueryHookOptions } from "@lib/ReactQuery"
-import { UserHandleError, UserHandle } from "@lib/users"
 import { useQuery, useQueryClient } from "@tanstack/react-query"
 import React, { useRef, useState } from "react"
 import { ActivityIndicator, Alert, StyleProp, ViewStyle } from "react-native"
