--- lib/Linkify.ts
@@ -1,13 +1,0 @@
-import linkifyIt from "linkify-it"
-import { linkifyAddUserHandleValidation } from "./users"
-import { linkifyAddEventHandleValidation } from "@event-details"
-
-/**
- * The app-wide `linkifyIt` config.
- */
-const linkify = linkifyIt()
-
-linkifyAddUserHandleValidation(linkify)
-linkifyAddEventHandleValidation(linkify)
-
-export { linkify }
